Don't reset uniq and uniqBy for each pipeline invocation
The current behavior is problematic as multiple steps invoke a
sub-pipeline that is executed for each item, it is pretty surprising
to see the uniqueness only apply for each item going into this
sub-pipelines. I think it is better that you need to make sure to get a
new invocation in each place you use uniqueness instead.